Turned the file size limit into a soft limit.
Record are now always written to the current file, even if that makes the file larger than the limit. If the file becomes too large, a new file is opened after the current record is written. This has two advantages: 1. the function WFile_storeRecordGzipCompressed becomes more efficient (since it no longer truncates the record it has just written); 2. it becomes possible to save records that are larger than the warc-max-size limit.

This behavior is similar to how the WarcWriter of Heritrix handles the file size limit.