Set the default WARC file size to infinite.
Wget will now by default create a single WARC file for the whole crawl.
Since most crawls will be smaller than 1GB (I think), this is a better
default: it avoids the unnecessary sequence number in the file name and
the metadata (log file, configuration) are written to the same WARC.
For small runs, this is what you want; for large runs, you'll probably
have a close look at the configuration anyway.