Fixed SEGV problem caused by logging for WARC and to a file.
log_vprintf_internal uses a variable argument list. In some specific
situations, when simultaneously writing to a log file and to the
temporary log file for the WARC, this argument list would be read
twice. This produced a segmentation fault.

Solution: generate the log string once and print that to the normal
log file and to the temporary WARC log. (Wget already did this for
log_dump_context.)