Try to handle redirects a bit better.
Redirects are somewhat difficult to resolve: Wget expects to receive
the file that a redirect points to *before* the redirect url itself.
In the WARC file, however, the redirect response often comes before
the file. Solution: keep a list of every redirect and its Location:
and resolve these at the end of the WARC file, when we have seen
every URL and file.